+++ src/core/api/AviationWeatherApi.ts
@@ -469,10 +469,10 @@ export class AviationWeatherApi {
                 typeof airport.freqs !== "string"
                     ? airport.freqs
                     : airport.freqs.split(";").map((f: string): AviationWeatherApiFrequency => {
-                          const parts = f.split(",");
+                          const [type, freq] = f.split(",");
                           return {
-                              type: parts[0],
-                              freq: parts[1] ? Number(parts[1]) : undefined,
+                              type: type ?? "",
+                              freq: freq ? Number(freq) : undefined,
                           };
                       }),
         };