+++ src/core/converter/parser/parseXml.ts
@@ -1,6 +1,6 @@
 export function parseXmlNode(xml: string, tag: string, defaultValue = ""): string {
-    const match = xml.match(new RegExp(`<${tag}[^>]*>(.*?)</${tag}>`, "ms"));
-    return match ? unXml(match[1]) : defaultValue;
+    const match = xml.match(new RegExp(`<${tag}[^>]*>(.*?)</${tag}>`, "ms"))?.[1];
+    return match ? unXml(match) : defaultValue;
 }
 
 export function parseXmlNodes(xml: string, tag: string, defaultValue: string[] = []): string[] {
@@ -11,14 +11,14 @@ export function parseXmlNodes(xml: string, tag: string, defaultValue: string[] =
 
 export function parseXmlAttribute(xml: string, attribute: string, defaultValue = ""): string {
     const regex = new RegExp(` ${attribute}="(.*?)"`, "ms");
-    const match = xml.match(regex);
-    return match ? unXml(match[1]) : defaultValue;
+    const match = xml.match(regex)?.[1];
+    return match ? unXml(match) : defaultValue;
 }
 
 function unXml(text: string): string {
-    const cdataMatch = text.match(/^<!\[CDATA\[(.+?)\]\]>$/);
+    const cdataMatch = text.match(/^<!\[CDATA\[(.+?)\]\]>$/)?.[1];
     return cdataMatch
-        ? cdataMatch[1]
+        ? cdataMatch
         : text.replace(/&([a-z]+);/g, (m, inner: string) => {
               switch (inner) {
                   case "lt":