+++ src/core/formatter/markdownTable.ts
@@ -10,6 +10,11 @@ export function markdownTable(rows: string[][]): string {
         });
     });
 
+    const alignmentRow = rows[1];
+    if (!alignmentRow) {
+        throw new Error("No alignment row found");
+    }
+
     return rows
         .map((row, rowIndex) => {
             const cells = row.map((cell, colIndex) => {
@@ -18,7 +23,7 @@ export function markdownTable(rows: string[][]): string {
                 }
 
                 // Center align
-                if (rows[1][colIndex]?.startsWith(":") && rows[1][colIndex]?.endsWith(":")) {
+                if (alignmentRow[colIndex]?.startsWith(":") && alignmentRow[colIndex]?.endsWith(":")) {
                     const padding = ((lengths[colIndex] ?? 2) - cell.length) / 2;
                     if (padding > 0) {
                         return " ".repeat(Math.floor(padding)) + cell + " ".repeat(Math.ceil(padding));
@@ -26,7 +31,7 @@ export function markdownTable(rows: string[][]): string {
                 }
 
                 // Left / right align
-                return rows[1][colIndex]?.endsWith(":")
+                return alignmentRow[colIndex]?.endsWith(":")
                     ? cell.padStart(lengths[colIndex] ?? 3, " ")
                     : cell.padEnd(lengths[colIndex] ?? 3, " ");
             });