+++ src/core/services/RoutePlanService.test.ts
@@ -12,14 +12,15 @@ describe("RoutePlanService", () => {
         const legs = routePlan.getRouteLegs();
 
         assert.strictEqual(flight.navigation.waypoints.length, legs.length + 1);
-        assert.strictEqual(flight.navigation.waypoints[0].identifier, legs[0].from);
-        assert.strictEqual(flight.navigation.waypoints[1].identifier, legs[0].to);
-        assert.strictEqual(legs[0].to, legs[1].from);
+        assert.strictEqual(flight.navigation.waypoints[0]?.identifier, legs[0]?.from);
+        assert.strictEqual(flight.navigation.waypoints[1]?.identifier, legs[0]?.to);
+        assert.strictEqual(legs[0]?.to, legs[1]?.from);
 
         //console.log(legs[0]);
-        assert.ok(75 < legs[0].track_deg && legs[0].track_deg < 76, "Tracking to KMTH");
-        assert.strictEqual(legs[0].wind_deg, flight.wind.directionInDegree, "Wind direction");
-        assert.ok(70 < legs[0].heading_deg && legs[0].heading_deg < 71, "Wind from the left, correction to the left");
+        const trackDeg = legs[0]?.track_deg ?? -1;
+        assert.ok(75 < trackDeg && trackDeg < 76, "Tracking to KMTH");
+        assert.strictEqual(legs[0]?.wind_deg, flight.wind.directionInDegree, "Wind direction");
+        assert.ok(70 < legs[0]?.heading_deg && legs[0]?.heading_deg < 71, "Wind from the left, correction to the left");
     });
 
     it("should calculate the legs and times with wind from 180°", () => {
@@ -30,14 +31,18 @@ describe("RoutePlanService", () => {
         const legs = routePlan.getRouteLegs();
 
         assert.strictEqual(flight.navigation.waypoints.length, legs.length + 1);
-        assert.strictEqual(flight.navigation.waypoints[0].identifier, legs[0].from);
-        assert.strictEqual(flight.navigation.waypoints[1].identifier, legs[0].to);
-        assert.strictEqual(legs[0].to, legs[1].from);
+        assert.strictEqual(flight.navigation.waypoints[0]?.identifier, legs[0]?.from);
+        assert.strictEqual(flight.navigation.waypoints[1]?.identifier, legs[0]?.to);
+        assert.strictEqual(legs[0]?.to, legs[1]?.from);
 
         //console.log(legs[0]);
-        assert.ok(75 < legs[0].track_deg && legs[0].track_deg < 76, "Tracking to KMTH");
-        assert.strictEqual(legs[0].wind_deg, flight.wind.directionInDegree, "Wind direction");
-        assert.ok(79 < legs[0].heading_deg && legs[0].heading_deg < 80, "Wind from the right, correction to the right");
+        const trackDeg = legs[0]?.track_deg ?? -1;
+        assert.ok(75 < trackDeg && trackDeg < 76, "Tracking to KMTH");
+        assert.strictEqual(legs[0]?.wind_deg, flight.wind.directionInDegree, "Wind direction");
+        assert.ok(
+            79 < legs[0]?.heading_deg && legs[0]?.heading_deg < 80,
+            "Wind from the right, correction to the right",
+        );
     });
 
     it("should calculate the legs and times with wind from 90°", () => {
@@ -50,13 +55,17 @@ describe("RoutePlanService", () => {
         //console.log(legs);
 
         assert.strictEqual(flight.navigation.waypoints.length, legs.length + 1);
-        assert.strictEqual(flight.navigation.waypoints[0].identifier, legs[0].from);
-        assert.strictEqual(flight.navigation.waypoints[1].identifier, legs[0].to);
-        assert.strictEqual(legs[0].to, legs[1].from);
+        assert.strictEqual(flight.navigation.waypoints[0]?.identifier, legs[0]?.from);
+        assert.strictEqual(flight.navigation.waypoints[1]?.identifier, legs[0]?.to);
+        assert.strictEqual(legs[0]?.to, legs[1]?.from);
 
         //console.log(legs[0]);
-        assert.ok(75 < legs[0].track_deg && legs[0].track_deg < 76, "Tracking to KMTH");
-        assert.strictEqual(legs[0].wind_deg, flight.wind.directionInDegree, "Wind direction");
-        assert.ok(76 < legs[0].heading_deg && legs[0].heading_deg < 77, "Wind from the front, correction to the right");
+        const trackDeg = legs[0]?.track_deg ?? -1;
+        assert.ok(75 < trackDeg && trackDeg < 76, "Tracking to KMTH");
+        assert.strictEqual(legs[0]?.wind_deg, flight.wind.directionInDegree, "Wind direction");
+        assert.ok(
+            76 < legs[0]?.heading_deg && legs[0]?.heading_deg < 77,
+            "Wind from the front, correction to the right",
+        );
     });
 });