+++ src/gui/renderer/AppState.ts
@@ -91,10 +91,11 @@ export class AppState {
     }
 
     protected formatDateTime(dateIn: Date): { date: string; time: string } {
-        const isoString = dateIn.toISOString().split("T");
-        const date = isoString[0];
-        const time = isoString[1].substring(0, 5);
-        return { date, time };
+        const [date, time] = dateIn.toISOString().split("T");
+        if (!date || !time) {
+            throw new Error(`Invalid date: ${dateIn}`);
+        }
+        return { date, time: time.substring(0, 5) };
     }
 
     protected getRoute() {